Add info on annotation needs
The documents added in this commit are derived from an earlier document
on the need for annotation support with links to use case documents.

The tooling document needs to be revised, as it now focuses on Web
Annotation clients and servers and uses tables, which are a nightmare to
maintain and extend in Markdown.